+++ functions/tsconfig.json
@@ -10,6 +10,7 @@
   },
   "compileOnSave": true,
   "include": [
-    "src"
+    "src",
+    "../src/app/interfaces"
   ]
 }
