+++ src/app/interfaces/note.ts
