+++ src/app/mypage/mypage-card/mypage-card.component.ts
@@ -3,8 +3,8 @@ import { NoteService } from 'src/app/services/note.service';
 import { UserService } from 'src/app/services/user.service';
 import { AuthService } from 'src/app/services/auth.service';
 
-import { Note } from 'functions/src/interfaces/note';
-import { User } from 'functions/src/interfaces/user';
+import { Note } from 'src/app/interfaces/note';
+import { User } from 'src/app/interfaces/user';
 
 import { Observable } from 'rxjs';
 import { ActivatedRoute } from '@angular/router';
