+++ src/app/shell/shell.component.ts
@@ -2,7 +2,7 @@ import { Component, OnInit } from '@angular/core';
 import { DrawerService } from '../services/drawer.service';
 import { AuthService } from '../services/auth.service';
 import { Observable } from 'rxjs';
-import { User } from '../../../functions/src/interfaces/user';
+import { User } from '../interfaces/user';
 
 @Component({
   selector: 'app-shell',
