+++ src/app/timeline/timeline-card/timeline-card.component.ts
@@ -1,5 +1,5 @@
 import { Component, OnInit, Input } from '@angular/core';
-import { NoteWithUser } from 'functions/src/interfaces/note';
+import { NoteWithUser } from 'src/app/interfaces/note';
 
 @Component({
   selector: 'app-timeline-card',
